+++ packages/coin-bitcoin/CHANGELOG.md
@@ -1,4 +1,3 @@
-
 # Change Log
 
 All notable changes to this project will be documented in this file.
@@ -7,7 +6,9 @@ All notable changes to this project will be documented in this file.
 
 ### Bug Fixes
 
-- **coin-bitcoin:** support network for psbtToSign ,support derivationPath derived path,add dustSize limit while building transaction,change some methods exportable ([498b582](https://github.com/okx/js-wallet-sdk/commit/498b5822ca7e9cc0de0dce52b8f14525cc47d330))
+- **coin-bitcoin:** support network for psbtToSign ,support derivationPath derived path,add dustSize limit while
+  building transaction,change some methods
+  exportable ([498b582](https://github.com/okx/js-wallet-sdk/commit/498b5822ca7e9cc0de0dce52b8f14525cc47d330))
 
 # [1.0.2](https://github.com/okx/js-wallet-sdk) (2023-11-24)
 
@@ -15,17 +16,30 @@ All notable changes to this project will be documented in this file.
 
 - **coin-bitcoin:** support btc nft/brc20 listing/buying ([34](https://github.com/okx/js-wallet-sdk/pull/34))
 
-
 # [1.0.3](https://github.com/okx/js-wallet-sdk) (2023-12-14)
 
 ### Feature
 
-- **coin-bitcoin:** support sighash type  'SIGHASH_ALL|SIGHASH_ANYONECANPAY' in psbt ([43](https://github.com/okx/js-wallet-sdk/pull/43))
-
+- **coin-bitcoin:** support sighash type  'SIGHASH_ALL|SIGHASH_ANYONECANPAY' in
+  psbt ([43](https://github.com/okx/js-wallet-sdk/pull/43))
 
 # [1.0.4](https://github.com/okx/js-wallet-sdk) (2023-12-21)
 
 ### Feature
 
-- **coin-bitcoin:** support two  memo types (string or hex ) by op_return  ([46](https://github.com/okx/js-wallet-sdk/pull/46))
+- **coin-bitcoin:** support two memo types (string or hex ) by
+  op_return  ([46](https://github.com/okx/js-wallet-sdk/pull/46))
+
+# [1.0.5](https://github.com/okx/js-wallet-sdk) (2023-12-27)
+
+### Feature
+
+- **coin-bitcoin:** support doginals  ([52](https://github.com/okx/js-wallet-sdk/pull/52))
+- **coin-bitcoin:** support src20 deploy,mint,transfer  ([58](https://github.com/okx/js-wallet-sdk/pull/58))
+
+# [1.0.6](https://github.com/okx/js-wallet-sdk) (2023-12-28)
+
+### Feature
+
+- **coin-bitcoin:** support psbt script-path spend  ([62](https://github.com/okx/js-wallet-sdk/pull/62))